add message page (/instructor/dashboard/message)
